Stop overriding models & controllers
## Original motivation

As far as I know, this mechanism was added so that it would be easier to
merge patches from the upstream repos. This is right: if we keep the
"core app" untouched, merging patches won't lead to merge conflicts.

However...

## Do we need it?

I don't think we want this because of some reasons:

* __We're not using it__. I don't see any flow of patches merged back
from upstream, it just seems like the repos tend to diverge and become
completely separate apps.

* __Gives a false sense of security__. The fact that merging patches
leads to no merge conflicts doesn't mean that the patches work. As a
matter of fact, because of the way this is implemented (not as an
engine, not as mixins, but just through class reopening), it seems to me
that it would be very likely to break if we were to merge patches from
the upstream repo. So getting some merge conflicts might actually be
beneficial because it will force whoever is porting patches to actually
review, try and evaluate if it makes sense to port them or not.

* __It's complicated__. The fact that every layer is run at two
different places is confusing, harder to grep, might lead to code
reloading issues because of the use of `require_dependency`, etc.

Maybe there are other reasons for doing this I haven't considered. If
that's the case, let's discuss those!